Add id to results table
This is a minor change to attempt to invalidate the CloudFront caches.

https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/Invalidation.html

I used Lambda@Edge to try to add security headers, but I am still
getting headers as would be served by the old Lambda version.

On adding security headers with Lambda@Edge: https://aws.amazon.com/blogs/networking-and-content-delivery/adding-http-security-headers-using-lambdaedge-and-amazon-cloudfront/